+++ hbc-core/src/main/java/com/twitter/hbc/RateTracker.java
@@ -13,174 +13,36 @@
 
 package com.twitter.hbc;
 
-import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Preconditions;
-
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ScheduledFuture;
-import java.util.concurrent.TimeUnit;
-
 /**
  * Tracks the rate of a recurring event with a sliding window.
  * Threadsafe
  */
-public class RateTracker {
+public interface RateTracker {
 
-  private final int granularityMillis;
-  private final int numBuckets;
-
-  private final RateUpdater rateUpdater;
-
-  private ScheduledFuture<?> future;
-  private ScheduledExecutorService executor;
-
-  public RateTracker(int granularityMillis, int numBuckets, boolean startPaused, ScheduledExecutorService executor) {
-    Preconditions.checkArgument(numBuckets > 0);
-    Preconditions.checkArgument(granularityMillis > 0);
-    Preconditions.checkArgument(granularityMillis / numBuckets > 0);
-
-    this.granularityMillis = granularityMillis;
-    this.numBuckets = numBuckets;
-    this.executor = Preconditions.checkNotNull(executor);
-
-    this.rateUpdater = new RateUpdater(startPaused);
-  }
-
-  public void eventObserved() {
-    rateUpdater.eventObserved();
-  }
+  public void eventObserved();
 
   /**
    * Pauses the rate tracker: the rate will be frozen.
    */
-  public void pause() {
-    rateUpdater.pause();
-  }
+  public void pause();
 
-  public void resume() {
-    rateUpdater.resume();
-  }
+  public void resume();
 
-  public void start() {
-    this.future = executor.scheduleAtFixedRate(rateUpdater, granularityMillis / numBuckets, granularityMillis / numBuckets, TimeUnit.MILLISECONDS);
-  }
+  public void start();
 
   /**
    * Stops tracking the rate
    */
-  public void stop() {
-    if (future != null) {
-      future.cancel(false);
-    }
-  }
+  public void stop();
 
   /**
    * Stops and shuts down the underlying executor
    */
-  public void shutdown() {
-    stop();
-    executor.shutdown();
-  }
-
-  /**
-   * Only used for testing
-   */
-  @VisibleForTesting
-  void recalculate() {
-    rateUpdater.run();
-  }
+  public void shutdown();
 
   /**
    * @return the current rate if it is available, NaN if not.
    * The rate is unavailable if <code>granularityMillis</code> hasn't elapsed
    */
-  public double getCurrentRateSeconds() {
-    return rateUpdater.getCurrentRateSeconds();
-  }
-
-  class RateUpdater implements Runnable {
-
-    private final int[] buckets;
-    private final Object lock;
-
-    private boolean paused;
-    private double rate;
-
-    private boolean rateValid;
-    private int total;
-    private int currentBucket;
-    private boolean previouslyPaused;
-    private int currentBucketCount;
-
-    RateUpdater() {
-      this(false);
-    }
-
-    RateUpdater(boolean paused) {
-      this.rateValid = false;
-      this.buckets = new int[numBuckets];
-      this.paused = paused;
-      this.rate = Double.NaN;
-      this.lock = new Object();
-    }
-
-    @Override
-    public void run() {
-      synchronized (lock) {
-        int currentCount = currentBucketCount;
-        currentBucketCount = 0;
-
-        if (paused) {
-          previouslyPaused = true;
-          return;
-        }
-
-        // skip the first estimation after a pause, since it could be in the middle of an estimation
-        if (previouslyPaused) {
-          previouslyPaused = false;
-          return;
-        }
-
-        int prevBucket = currentBucket;
-        currentBucket = (currentBucket + 1) % numBuckets;
-
-        if (currentBucket == 0) {
-          // we've wrapped around again. this rate is now valid
-          rateValid = true;
-        }
-
-        int prevBucketCount = buckets[prevBucket];
-        buckets[prevBucket] = currentCount;
-
-        total += currentCount - prevBucketCount;
-        if (rateValid) {
-          rate = total * 1000d / granularityMillis;
-        }
-      }
-    }
-
-    public void eventObserved() {
-      synchronized (lock) {
-        currentBucketCount++;
-      }
-    }
-
-    public void pause() {
-      synchronized (lock) {
-        paused = true;
-      }
-    }
-
-    public void resume() {
-      synchronized (lock) {
-        paused = false;
-      }
-    }
-
-    public double getCurrentRateSeconds() {
-      synchronized (lock) {
-        return rate;
-      }
-    }
-  }
+  public double getCurrentRateSeconds();
 }
